+++ src/main/java/com/yoedu/yoedurealestateapi/domain/entities/PropertyType.java
@@ -31,7 +31,7 @@ public class PropertyType {
     private String icon;
 
     @Column(name = "sort_order", nullable = false)
-    private Short sortOrder = 0;
+    private Integer sortOrder = 0;
 
     @Column(name = "is_active", nullable = false)
     private Boolean isActive = true;
